moved around some query tests and fixed other failing tests
